get it to the point where I can open positions and view them
